backend/requirements.txt
fastapi
hypercorn
pandas
numpy
pydantic
joblib
catboost; python_version < "3.13"
python-dotenv

